pyproject.toml
[project]
name = "phone-caller-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.1",
    "langchain>=0.3.26",
    "langchain-gigachat>=0.3.11",
    "langgraph>=0.5.0",
    "langgraph-cli[inmem]>=0.3.3",
    "python-dotenv>=1.1.1",
]
